requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==23.2.1

# PDF processing
pymupdf==1.24.10

# Translation — Google Translate fallback (always available, no key needed)
deep-translator==1.11.4

# Translation — LLM backends (install the one matching your LLM_PROVIDER)
google-generativeai>=0.7.0   # for LLM_PROVIDER=gemini
openai>=1.0.0                 # for LLM_PROVIDER=openai | custom

# Utilities
beautifulsoup4==4.12.3
